requirements.txt
aiohttp>=3.9.0
anthropic>=0.25.0
click>=8.1.7
colorlog>=6.8.2
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dateutil>=2.9.0
PyYAML>=6.0.1
rich>=13.7.0
tenacity>=8.2.3
tzdata>=2024.1
